[ftl] Fixed the logger
During testing, I noticed that the logger doesn't work properly when
attaching to an existing FTL volume (i.e. when not formatting): the FTL
ended up using the stderr logger which doesn't work within a test. I
found the initialisation of the logger confusing, so I've restructured
things so that it's harder to get it wrong.

Change-Id: Ic3cf64775a1fef1214904839fc40cf2cf5e84c33
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/528790
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Chris Suter <csuter@google.com>