[fidl][llcpp] Error logging support
We would like to hide `fidl::Reason` eventually from results and unbind
info. This implies that users shouldn't fetch and log the raw
`fidl::Reason` enum, which only results in a cryptic number anyways.

As a step towards that, this CL adds logging support to `fidl::Result`
and `fidl::UnbindInfo`:

- They can be logged to output streams, which includes FX_LOGS, gtest
assertions, `std::cout`/`std::cerr`, and so on.
- They can also be logged with `printf`-like formatters. However, this
is slightly less performant compared to the streams API, since we need
to build a temporary string on the heap first.

Based on discussions with the Diagnostics team, logging APIs are moving
to a structured-K/V pair model, but the structured API doesn't support
custom user types yet. When fxbug.dev/76211 is ready, we would be able
to make use of that. For now, string-based logging is the next best
thing.

An example of logging a result to syslog:

  fidl::WireResult bar = fidl::WireCall(foo_client_end).GetBar();
  if (!bar.ok()) {
    FX_LOGS(ERROR) << "GetBar failed: " << bar.error();
  }

When an error happens, the rendered log would look along the following lines:

  GetBar failed: FIDL operation failed due to decode error, status: ZX_ERR_INVALID_ARGS (-10), detail: non-nullable string was absent

Another example for when a `fidl::Client` unbinds:

  client/server endpoint was unbound due to peer closed, epitaph: ZX_ERR_INTERNAL (-1)

Bug: 75702
Change-Id: Idf4772b61487c8b7376292a8fe6191979ead8919
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525500
Commit-Queue: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>