[zxio] Move zircon/system/ulib/zxio to sdk/lib/zxio
This moves the contents of the directory:
  zircon/system/ulib/zxio
to the directory:
  sdk/lib/zxio

This commit includes the following forwarding build target to ease the
transition to the new layout:

  //zircon/system/ulib/zxio:zxio

As this target is used by some rules outside the fuchsia git repository.

This commit is generated by the script //scripts/move_source/move_source.py and
should contain no behavioral changes outside a bug fix for the script
when trying to move files that have matches in checked-in binary files.

Several allowlists reference zxio and are also updated to refer to the
new location.

Bug: 36063
Change-Id: Iea2702bf08871acb6e9065dc402b077501ec9f43
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533021
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>