[audio][aml-g12-tdm] Allow multiple mclk pad usage
When using multiple audio engines as in TDM_B for speaker output
and TDM_A for BT output, we need to allow for different MCLKs to
be enabled simultaneously. With this CL we won't override the MCLK
configuration of one engine with the mclk configuration of another.

Fixed: 75872
Test: aml-g12-tdm-test
Test: Manual audio check on Astro and Sherlock
Change-Id: Ifc443832f3a3028057da9c5ef2ec0f5585248478
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532064
Commit-Queue: Andres Oportus <andresoportus@google.com>
Reviewed-by: Eric Holland <hollande@google.com>