[ftl] Reject map pages with all 0xff in the second half of OOB.
This is a reland that adds the changes to the ftl-mtd to override the
NdmBaseDriver methods for supplying page and spare sizes. Original
change was fxr/529803

This rejects map pages with all 0xff in the second half of the OOB in
order to reject partial page writes which result in wear count listed as
the max value and the validity bits are 0xff, though validity bits are
currently all 0xff.

This also starts writing a form of checksum to the last validity byte to
ensure that the write completed, and to spot any unexpected 0xff
patterns in the oob that may have resulted from an incomplete write.

Bug: b/179400686
Change-Id: If280d160cf0cb4c3917f2dcf897df5cb76a15476
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534962
Reviewed-by: Vikram Auradkar <auradkar@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>