[x86][acpi] Publish all devices in the ACPI tables.
This CL introduces a new acpi::Manager class, which is responsible for
the three phases of device discovery, bus type inference, and
publishing.

For now, these devices have no bind properties associated with them.

Currently, the old code in acpi-nswalk.cc is left intact. We will remove
it as drivers move out of the board driver.

Additionally, some basic infrastructure for testing ACPI functionality is
added, including a mock of (some) of the functionality exposed in the
Acpi class. Currently we just have a few tests that verify that ACPI is
interacted with correctly and all the devices we expect to be published
are published.

Bug: 78348
Change-Id: I65be9691fcf8113091e8448d80bf41c1a10f4983
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/543309
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>