[pci] Support making PCI devices composite with ACPI.
This CL adds a new list of PciBdf to PciPlatformInfo. For any device in
this list, we add a third composite fragment to the composite devices
which binds to the ACPI device that represents the device at that BDF.

For now, nothing populates this list, and so no devices are composite
with ACPI. A future CL will populate this list in the x86 board driver
and start creating composites.

Bug: 78565
Change-Id: I89c08593140030bde32f6b86be35146ddcd4e40a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545150
API-Review: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>