[usb] Reduce logging severity from aml_usb_phy_v2 driver
If the phy_rdy register is not set within a 1ms interval, the driver
currently logs to ERROR level, but otherwise continues normally. As this
does not appear to be a fatal condition, the log is reduced to WARNING
level. Other public drivers for this same hardware do not check the
phy_rdy register.

Bug: 79564
Change-Id: I5499655b2b963c5cb41986d11037236b7069ff15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548351
Commit-Queue: Sam Hansen <hansens@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>