[fidlc] Introduce NamingContext for anonymous names
This CL adds a new variant of flat::Name that represents
anonymous names, and uses it to properly handle such names.
For example, some behavior changes include:

* anonymous names will always have a naming context (the
  list of names) associated with them
* they will also have a flattened name generated based on
  this context, which will be guaranteed to be unique (no
  more SomeLongAnonymousPrefixN)
* users cannot reference such flattened names in FIDL
  source

The bulk of the logic for handling anonymous names is
in the NamingContext class.

This CL only updates the uses of anonymous names that
existed before the new syntax, i.e. top level request/response
names, as well as generated result type names. The rest
of the changes are separated into another CL.

Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler
Test: fx test fidlc_golden_tests
Bug: 73285
Bug: 75112

Change-Id: Ie402273f59f0f360d54ab2549745c5035e6624df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545841
Commit-Queue: Felix Zhu <fcz@google.com>
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>