[driver-index] Use driver-components
This CL updates the driver-index to read a driver component
manifest in order to find where the driver library and the
bind program are located.

Once driver-index is included in the build, DriverManager will
use driver-index to find and bind packages that live in base
drivers.

The ddk-firmware-test goes through the whole process of
DriverManager interacting with DriverIndex to load a driver
from a package.

Fixed: 78958
TEST: ddk-firmware-test, driver-index-unittests

Change-Id: I2b02a601211320f71ce8044a87c81712f7a27ce0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546906
Commit-Queue: David Gilhooley <dgilhooley@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>