[fdf][composite] Add primary_fragment field.
To align with the future change, where one of the fragments will be
identified as the "primary" fragment. A future change will replace the
`coresident_device_index` with a boolean which describes whether or not
to load the driver which binds to the composite device in the same
driver host as the one hosting the primary fragment.

Some of primary fragment choices are a bit awkwards, and a // ???
comment is present to denote them.

Change-Id: I283bce7a7f9a0acaee95920be4a6d95702b6cc1b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551896
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>