[fdf] Add option to reboot on driver crashes
For products where we aren't confident in the system's ability to
recover from driver crashes, we would like to reboot the system instead
when we observe a crash.

The currently implementation just crashes the driver manager in response
to a driver host crashing, which will trigger the system to reboot due
to the fact driver manager is marked as a critical component. This can
be updated to do something more graceful in the future.

devmgr-integration-test was updated to support exposing the process
handle for driver manager, as well as share the exception info with the
exception callback registered.

Bug: 67168
Tested: fx test isolated-devmgr-tests
Change-Id: I8749261dfe5f70b81885051e06455010a4fa824f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/552990
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>