[x86][acpi] Add support for Device Tree namespace link device ID.
This CL adds support for the "Device Tree namespace link device ID."
Essentially, if a device has a _HID or _CID value of "PRP0001", we
ignore that value and instead use the value of the "compatible" property
in the _DSD device properties.

See
https://www.kernel.org/doc/html/latest/firmware-guide/acpi/enumeration.html#device-tree-namespace-link-device-id
for more information.

Bug: 78348
Bug: 76095
Change-Id: I0b702d0fcf65c4ee28fce0ac55b2a7f5f3d5e79e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/553861
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>