[driver_manager] Fix unbind / remove task reference errors.
When SendUnbind or SendCompleteRemove failed, the device would
take ownership of the completion, while the task would try
to call the (now moved) completion.

There was also another ref_ptr issue when SendRemove failed.
The RemoveTask would ask the coordinator to forcibly remove
the device, which would drop the last reference to the device
before the RemoveTask completed. This would trigger a ref_ptr
assert when the task tried to take a new reference in Task::Complete.

Change-Id: I01bd312e1c005e1201e93fccc7bafdffd827945b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/560144
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>