[build] Introduce packaged_shared_library() template.
Add //build/dist/packaged_shared_library.gni which is essentially
a rewrite of variant_shared_library_redirect(), and deprecate the
latter (since it is still used by //vendor/... build rules).

The template is used to guarantee that a shared_library() instance
is always installed into a Fuchsia package, even if it is not an
actual ELF runtime dependency of one of the final executable or
loadable module targets in the package.

Bug: 82489

Change-Id: Ia767edd7732a1b3216cde5d064cacf6b139bd619
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567764
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: David Fang <fangism@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>