[pci] Handle MSI-X configurations sharing BARs
MSI-X capabilities can specify that their supporting tables may exist in
up to two device BARs, and that they may overlap with a BAR that a
device uses for data. Previously, we simply denied access to the BAR.
But now as long as we can ensure that the vector table and pending bit
array do not share a page with device data we can provide a mapping of
the space with that information excluded.

Additionally, a bug was found where inited_ was not being set.  This has
been fixed. Tests have been added both for Init as well as the new MSI-X
support.

Bug: 32978, 80426
Change-Id: I72bb46e83584b3aefde3f3dc4799195fa87379d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569223
Commit-Queue: Christopher Anderson <cja@google.com>
Fuchsia-Auto-Submit: Christopher Anderson <cja@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>