[storage][tests[: Run in isolation.
The previous situation would keep the component(isolated devmgr) as part
of a group of packages that contained the storage tests. This means,
that if the user added the group, the test dependencies couldbe
resolved.

Now this change, allows a user just adding package A(not the group), and
have this package contain all the components it needs to execute
successfully.

in fx args the situation would be:

Before: universe -> [my test, isolated-devmgr-component]
Now: universe -> [my test]

Test: fx test fs-managemente-tests (and CQ)
Change-Id: I55032532436406ba1377d40f9e520c024450fda0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/571445
Fuchsia-Auto-Submit: Gianfranco Valentino <gevalentino@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Simon Shields <simonshields@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>
Reviewed-by: Chris Suter <csuter@google.com>