[llcpp] Use move semantics for LLCPP resource unions
After the move to the v2 wire format, union values with payloads <= 4
bytes will be inlined on to the union. For the case of handles, this
means that the handle value will be stored on the union rather than
having a pointer to the handle.

A consequence of this is that unions with single handles (sent alone
or in wrapping structs / arrays) will no longer be able to be
copy constructed or copy assigned. For the sake of providing a
consistent API, this CL makes all resource unions move-only.

Bug: 79578
Change-Id: I772e49f9c05d9107bf5908f6d85852499e35f209
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/577261
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>