[usb-hub] Disconnect all devices on transaction failure.
Previously, when an entire USB hub was disconnected the hub device would
be correctly destroyed, but it would not tear down all of its child
devices, and it would print out an error message "Could not get port
status".

The error message was because the interrupt callback didn't correctly
propogate the response status up to the interrupt thread, which would
cause the thread to sometimes try handling an interrupt that hadn't
happened.

To properly disconnect all USB devices, we disconnect any port that was
previously in the "connected" state after the interrupt loop has
finished (either because unbind() was called or some other error
occurred, e.g. on disconnect).

Multiply: usb-hub-package
Bug: 78415
Change-Id: Iad852567a41cfb02a911bfbd75b10aaa4f501f94
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579761
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Sam Hansen <hansens@google.com>