[aml-pwl-regulator] Copy fields of FIDL table to avoid lifetime bug
The existing AmlPwmRegulator is constructed by moving a FIDL table
object into the AmlPwmRegulator. However, this move only moves the
top-level table header and not the contents of the table, which
only lives until the fidl::DecodedMessage object goes out of
scope. Therefore, accesses to the vreg_range_ field after Create()
exits are use-after-free.

This CL fixes this issue by copying the contents of the table
field-by-field to the AmlPwmRegulator object.

Change-Id: I9c4cd2cabe06e3abc913a6d7d5848e3d5ab7e6fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581701
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>