[sherlock][i2c] Publish LCD bias module for testing.
The LCD bias module is not configured by the operating system so it was
never published by the i2c driver. We need to probe this module for
testing so we expose it on the bus now.

Test: device-enumeration-test
i2cutil r /dev/class/i2c/007 0x03

Change-Id: Ida32fb43fdd6b4b93f18f4fa82c2953c03ccf452
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584156
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: Gurjant Kalsi <gkalsi@google.com>