[fidl][llcpp] Fail TryDispatch early for invalid message
WireTryDispatch tries to dispatch the message against a server
implementation, and doesn't consume the message if a handler is not
found. This is used to support ad-hoc composing several protocols by
trying to dispatch one after the other in a chain of TryDispatch.

This CL is a collection of some improvements:

 - If the message has an error, exit the chain of TryDispatch early
 since it is meaningless to keep trying otherwise.
 - Mark the return value [[nodiscard]] which also revealed one usage
 that forgot ignored the return value.

It would be ideal if we could remove TryDispatch since it subverts the
FIDL protocol semantics. However, there are still a few drivers that are
load-bearing on it. See fxbug.dev/85473.

Bug: 85473
Test: fx test -v //src/lib/fidl/llcpp

Change-Id: I9366dd5a31f829daa65be16276ef10937125318d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/586486
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
API-Review: Ian McKellar <ianloic@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>