[x86][acpi] Refactor ownership a bit.
Previously, ACPI was initialised as follows:

1. The X86 driver proper creates an acpi::Acpi instance, which it passes
   to publish_acpi_devices().
2. publish_acpi_devices() creates an acpi::Manager on the stack, which
   does device discovery and publishing.
3. Each ACPI device was given a pointer to the Acpi implementation, and
   created its own async loop (and thread) to handle FIDL requests.

In these circumstances, the Manager is destroyed as soon as all devices
are published.

In preparation for upcoming ACPI event work, this CL refactors the
ownership of various objects to be like this:

1. The X86 driver creates an acpi::Acpi and an acpi::Manager. It passes
   the acpi::Manager to publish_acpi_devices(). The acpi::Manager
   contains a single async::Loop for all ACPI devices.
2. publish_acpi_devices call the discovery/publish methods.
3. Each ACPI device is given a pointer to the acpi::Manager.

This means that we have less FIDL threads, and that future work that
involves sharing state between multiple devices can be done in the
acpi::Manager.

Bug: 85785
Change-Id: Ic597d72e1f8d492a3955582e6521b759be0bf272
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588601
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>