[rust] add allows for holding over await point
Some types in the new version of rustc are marked as being
unsafe to hold over an await point.  These types are marked with
a must_not_suspend lint and produce a warning.

Allow the current usages in the tree explicitly, with a pointer to
the bug for a further explanation and next steps.

Bug: 86245
Change-Id: If46d05d46af267e08c04911b7c657f22d4607bc0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591287
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Commit-Queue: Joseph Ryan <josephry@google.com>