[llcpp] Simplify unique_ptr/optional<WireSyncClient>
In the early history of `MyProtocol::SyncClient`, the class did not have
a default constructor in order to force the user to provide a channel.
This complicated control flows where the client is lazily initialized:
users either have to wrap the client in a unique_ptr or an optional, and
populate it later.

Today, `WireSyncClient` already provides a default constructor. The
unique_ptr and optional wrappers don't add too much value - the client
itself could be default constructed with an invalid channel, and later
replaced with a valid channel. Furthermore, from
https://groups.google.com/a/fuchsia.dev/g/fidl-dev/c/nF72xpPL4gU we are
planning to switch sync clients to the "->" operator to make calls. This
doesn't interact well with clients wrapped in unique_ptr/optional,
because ones now has to write `(*client)->Foo()` instead of
`client->Foo()`.

This CL explores converting all instances of unique_ptr/optional wrapped
sync clients to default constructed sync clients.

In terms of safety, the new version does not reduce safety, since
neither unique_ptr nor optional check for validity on dereference.
Making a call on an absent client would lead to undefined behavior, and
most uses of unique_ptr/optional are not checking for presence.
Making a call on an invalid client crashes with a consistent message.

In terms of performance, the new version may improve performance since
now there are less boxing and indirections.

Change-Id: Ie147fd6c4dd2577589ff5d631370bd44ca97d31f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592844
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Yilong Li <liyl@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>