[drivers][nand] Add stats for bit flips and read failures.
Adding inspect metrics for read failures and ECC bit flip counts. This
will be consumed by Sampler for export to cobalt to build reports on the
distribution of attempts and recovered bit flips via the histogram, as
well distribution of count of internal and complete failure per hour or
day.

Bug: 85321

Change-Id: Ic493401c5927be2dfeadc06e2b5fa5681a03c037
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593101
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Chris Suter <csuter@google.com>